+++ Sources/Chartdesk/Views/MacSearchField.swift
@@ -3,7 +3,7 @@ import SwiftUI
 
 extension Notification.Name {
     static let showPerformance = Notification.Name("ChartdeskShowPerformance")
-    static let showWeather = Notification.Name("ChartdeskShowWeather")
+    static let focusWeather = Notification.Name("ChartdeskFocusWeather")
     static let focusAirportSearch = Notification.Name("ChartdeskFocusAirportSearch")
     static let focusChartSearch = Notification.Name("ChartdeskFocusChartSearch")
 }